move compcor to confounds.py
